--- csrc/npu/CachingAllocatorHelper.h
@@ -1,137 +1,0 @@
-#include <c10/core/Device.h>
-#include <c10/util/irange.h>
-namespace c10_npu::NPUCachingAllocator {
-
-//
-// Struct and factory of ExpandableSegment
-//
-
-struct SegmentRange {
-  char* ptr;
-  size_t size;
-  SegmentRange(void* p, size_t s) : ptr(static_cast<char*>(p)), size(s) {}
-};
-
-/*
-Note [Expandable Segments]
-
-Rationale
-
-For large (>2MB) allocations, the allocator calls cudaMalloc to get allocations
-that are the same size as what the user requests. In the future, parts of these
-allocations can be reused for other requests if they are free. This works well
-when the program makes many requests of exactly the same size or of sizes that
-even multiples of that size. Many deep learning models follow this behavior.
-However, one common exception is when the batch size changes slightly from one
-iteration to the next, e.g. in batched inference. When the program runs
-initially with batch size N, it will make allocations appropriate for that size.
-If in the future, it runs at size N - 1, the existing allocations will still be
-big enough. However, if it runs at size N + 1, then it will have to make new
-allocations that are slightly larger. Not all the tensors are the same size.
-Some might be (N + 1)*A and others (N + 1)*A*B where A and B are some non-batch
-dimensions in the model. Because the allocator reuses existing allocations when
-they are big enough, some number of (N + 1)*A allocations will actually fit in
-the already existing N*B*A segments, though not perfectly. As the model runs it
-will partially fill up all of these segments leaving unusable free slices of
-memory at the end of these segments. The allocator at some point will need to
-cudaMalloc a new (N + 1)*A*B segment. If there is not enough memory, there is
-now no way to recover the slices of memory that are free at the end of existing
-segments. With models 50+ layers deep, this pattern might repeat 50+ times
-creating many slivers.
-
-Approach
-
-Expandable segments allows the allocator to create a segment initially and then
-expand its size later when more memory is needed. Instead of making one segment
-per allocation, it tries to make one segment (per stream) that grows as
-necessary. Now when the N + 1 case runs, the allocations will tile nicely into
-the one large segment until it fills up. Then more memory is requested and
-appended to the end of the segment. This process does not create as many slivers
-of unusable memory, so it is more likely to succeed at finding this memory.
-
-Implementation
-
-The expandable_segments:True option is used to enable/disable this behavior. We
-use cuda's low-level memory APIs, which are similar to mmap, to extend the
-memory segments. These APIs separate the allocation of physical memory
-(cuMemCreate) from the allocation of virtual address space (cuMemAddressReserve)
-and the associate between them cuMemMap/cuMemSetAccess.
-
-When we allocate a new segment, we allocate enough address space to map
-basically the entire physical memory of the GPU (there is 256TiB of address
-space), but we only map enough physical memory to handle the current amount of
-memory needed by the program. As more is requested, we add more physical memory
-to the segment. This can work at the granularity of GPU pages which are 2MiB
-currently.
-
-If we end up out of memory, we can unmap all the memory in our segment
-corresponding to empty physical pages, and return it to CUDA for use at another
-address in the segment or in a segment for a different stream.
-
-A current limitation of CUDA's API is that physical memory
-(CUmemGenericAllocationHandle) cannot be split up after it is mapped even if the
-handle holds multiple GPU pages. The cost to map/unmap memory is proportional to
-the number of physical memory chunks that were allocated (mapping 10 separately
-allocated 2MiB pages takes 10x time compared to mapping one 20MiB physical
-allocation of 10 pages).  Changing memory mappings also appears to involve at
-least some synchronous actions with the GPU and so should be considered an
-expensive operation. To limit overhead, we use 2MiB pages for our small pool and
-20MiB pages for our large pool. Initially allocation using expandable_blocks
-will be slower than cudaMalloc, though still in the milliseconds range for
-mapping the entire memory.
-
-When mapping new memory to expand the segment, we look for the lowest address at
-which we can fit a new allocation by adding new pages. Normally this will be at
-the end of the block. But if have previously unmapped blocks earlier in the
-segment during an OOM, it will first try to fill in those gaps to keep the
-segment as a single block. By allocating at the lowest address we encourage
-the split up parts of the block to merge into a single block again, reducing
-fragmentation potential.
-
-Allocation of blocks in the segment uses the same best-fit heuristics of the
-rest of the allocator.
-
-Expandable blocks can be enabled/disabled throughout the run of a program. When
-disabled, the allocator will not put new allocations in an expandable block.
-
-Limitations
-
-* Slightly slower initial memory allocation speed.
-* IPC of cuda tensors (e.g. for multiprocess dataloaders) is not supported.
-However, it is possible to temporarily disable (expandable_segments:False) the
-bevhavior for allocator tensors that need to be used cross-process.
-* CUDA runtime APIs related to sharing memory across process
-(cudaDeviceEnablePeerAccess) do not work for memory allocated with cuMemMap.
-Instead these mapping have to be done manually. The allocator now has an
-`enablePeerAccess` method to do this.
-*/
-class ExpandableSegment {
- public:
-  // begin must be aligned to segment_size_.
-  // returns the actual range mapped, which may be
-  // greater than requested if size is not aligned to segment_size_.
-  // return size of 0 indicates OOM
-  virtual SegmentRange map(SegmentRange range) = 0;
-  // unmaps all the completely empty segment_size_ segments between
-  // [begin, begin + size), returns the offset where the range begin,
-  // and the actual size unmapped (multiple of segment_size_)
-  virtual SegmentRange unmap(SegmentRange range) = 0;
-  // segment size
-  virtual size_t size() const = 0;
-  // ptr of memory address
-  virtual char* ptr() const = 0;
-  virtual ~ExpandableSegment() = default;
-};
-
-ExpandableSegment* createExpandableSegment(
-    int device,
-    void* stream,
-    size_t size);
-
-// Wraps the insert event function
-void insertEventWrapper(c10::DeviceIndex device, std::function<void()> insertEventFn);
-
-// Returns the current stream for the given device
-void* getCurrentStream(c10::DeviceIndex);
-
-} // namespace c10_npu::NPUCachingAllocator
