+++ npu/aten/common/TensorFactories.cpp
@@ -122,7 +122,7 @@ at::Tensor NPUNativeFunctions::empty(
       OPS_ERROR(ErrCode::NOT_SUPPORT));
   check_size_nonnegative(size);
   c10_npu::NPUGuard guard_(device_);
-  c10::Allocator* allocator = c10_npu::NPUCachingAllocator::get();
+  c10::Allocator* allocator = c10_backend::CachingAllocator::NPU::get();
   int64_t nelements = c10::multiply_integers(size);
   auto dtype = c10::scalarTypeToTypeMeta(dtype_or_default(dtype_opt));
   int64_t size_bytes = nelements * dtype.itemsize();
@@ -313,7 +313,7 @@ at::Tensor NPUNativeFunctions::empty_with_format(
       OPS_ERROR(ErrCode::NOT_SUPPORT));
   check_size_nonnegative(size);
   c10_npu::NPUGuard guard_(device_);
-  c10::Allocator* allocator = c10_npu::NPUCachingAllocator::get();
+  c10::Allocator* allocator = c10_backend::CachingAllocator::NPU::get();
   // when the shape and format are not match, fix format here.
   aclFormat format =
       InferFormat::GuessStorageFormat(size, (aclFormat)dst_format);
