+++ npu/aten/common/from_blob.cpp
@@ -38,7 +38,7 @@ at::Tensor TensorMaker::make_tensor() {
       dtype_or_default(c10::optTypeMetaToScalarType(opts_.dtype_opt())));
   at_npu::native::check_size_nonnegative(sizes_);
   c10_npu::NPUGuard guard(*device_);
-  c10::Allocator* allocator = c10_npu::NPUCachingAllocator::get();
+  c10::Allocator* allocator = c10_backend::CachingAllocator::NPU::get();
 
   std::size_t size_bytes = computeStorageSize();
