+++ torch_npu/csrc/npu/Module.cpp
@@ -63,7 +63,7 @@ void RegisterNPUDeviceProperties(PyObject* module) {
           size_t)>(torch_npu::_record_memory_history));
 
   m.def("_npu_isHistoryEnabled", []() {
-    return c10_npu::NPUCachingAllocator::isHistoryEnabled();
+    return c10_backend::CachingAllocator::NPU::isHistoryEnabled();
   });
 }
 
@@ -123,8 +123,8 @@ void RegisterNpuPluggableAllocator(PyObject* module) {
   auto m = py::handle(module).cast<py::module>();
 
   py::class_<
-      c10_npu::NPUCachingAllocator::NPUAllocator,
-      std::shared_ptr<c10_npu::NPUCachingAllocator::NPUAllocator>>(
+      c10_backend::CachingAllocator::CachingAllocator,
+      std::shared_ptr<c10_backend::CachingAllocator::CachingAllocator>>(
       m, "_npu_NPUAllocator");
   m.def("_npu_getAllocator", []() {
     return py::cast(torch::npu::NPUPluggableAllocator::getCurrentAllocator());
@@ -132,13 +132,13 @@ void RegisterNpuPluggableAllocator(PyObject* module) {
 
   m.def(
       "_npu_changeCurrentAllocator",
-      [](std::shared_ptr<c10_npu::NPUCachingAllocator::NPUAllocator>
+      [](std::shared_ptr<c10_backend::CachingAllocator::CachingAllocator>
              allocator) {
         torch::npu::NPUPluggableAllocator::changeCurrentAllocator(allocator);
       });
   py::class_<
       torch::npu::NPUPluggableAllocator::NPUPluggableAllocator,
-      c10_npu::NPUCachingAllocator::NPUAllocator,
+      c10_backend::CachingAllocator::CachingAllocator,
       std::shared_ptr<
           torch::npu::NPUPluggableAllocator::NPUPluggableAllocator>>(
       m, "_NPUPluggableAllocator")
@@ -422,14 +422,14 @@ PyObject* THNPModule_setMemoryFraction(PyObject* _unused, PyObject* args) {
   double fraction = PyFloat_AsDouble(fraction_o);
   int64_t device = PyLong_AsLongLong(device_o);
 
-  c10_npu::NPUCachingAllocator::setMemoryFraction(fraction, device);
+  c10_backend::CachingAllocator::NPU::setMemoryFraction(fraction, device);
   END_HANDLE_TH_ERRORS
   Py_RETURN_NONE;
 }
 
 PyObject* THNPModule_emptyCache(PyObject* _unused, PyObject* noargs) {
   HANDLE_TH_ERRORS
-  c10_npu::NPUCachingAllocator::emptyCache();
+  c10_backend::CachingAllocator::NPU::emptyCache();
   END_HANDLE_TH_ERRORS
   Py_RETURN_NONE;
 }
@@ -442,10 +442,10 @@ PyObject* THNPModule_memoryStats(PyObject* _unused, PyObject* arg) {
       PTA_ERROR(ErrCode::PARAM));
   const int device = (int)THPUtils_unpackLong(arg);
 
-  using c10_npu::NPUCachingAllocator::DeviceStats;
-  using c10_npu::NPUCachingAllocator::Stat;
-  using c10_npu::NPUCachingAllocator::StatArray;
-  using c10_npu::NPUCachingAllocator::StatType;
+  using c10_backend::CachingAllocator::DeviceStats;
+  using c10_backend::CachingAllocator::Stat;
+  using c10_backend::CachingAllocator::StatArray;
+  using c10_backend::CachingAllocator::StatType;
 
   const auto statToDict = [](const Stat& stat) {
     py::dict dict;
@@ -468,7 +468,7 @@ PyObject* THNPModule_memoryStats(PyObject* _unused, PyObject* arg) {
   };
 
   const DeviceStats stats =
-      c10_npu::NPUCachingAllocator::getDeviceStats(device);
+      c10_backend::CachingAllocator::NPU::getDeviceStats(device);
 
   py::dict result;
   result["num_alloc_retries"] = stats.num_alloc_retries;
@@ -498,7 +498,7 @@ PyObject* THNPModule_resetAccumulatedMemoryStats(
       "invalid argument to reset_accumulated_memory_stats",
       PTA_ERROR(ErrCode::PARAM));
   const int device = (int)THPUtils_unpackLong(arg);
-  c10_npu::NPUCachingAllocator::resetAccumulatedStats(device);
+  c10_backend::CachingAllocator::NPU::resetAccumulatedStats(device);
   END_HANDLE_TH_ERRORS
   Py_RETURN_NONE;
 }
@@ -510,7 +510,7 @@ PyObject* THNPModule_resetPeakMemoryStats(PyObject* _unused, PyObject* arg) {
       "invalid argument to reset_peak_memory_stats",
       PTA_ERROR(ErrCode::PARAM));
   const int device = (int)THPUtils_unpackLong(arg);
-  c10_npu::NPUCachingAllocator::resetPeakStats(device);
+  c10_backend::CachingAllocator::NPU::resetPeakStats(device);
   END_HANDLE_TH_ERRORS
   Py_RETURN_NONE;
 }
@@ -529,8 +529,8 @@ torch::CapturedTraceback* getFromContext(
 PyObject* THNPModule_memorySnapshot(PyObject* _unused, PyObject* noargs) {
   HANDLE_TH_ERRORS
 
-  using c10_npu::NPUCachingAllocator::BlockInfo;
-  using c10_npu::NPUCachingAllocator::SegmentInfo;
+  using c10_backend::CachingAllocator::BlockInfo;
+  using c10_backend::CachingAllocator::SegmentInfo;
 
   py::str device_s = "device";
   py::str address_s = "address";
@@ -603,7 +603,7 @@ PyObject* THNPModule_memorySnapshot(PyObject* _unused, PyObject* noargs) {
     return segmentDict;
   };
 
-  auto snapshot = c10_npu::NPUCachingAllocator::snapshot();
+  auto snapshot = c10_backend::CachingAllocator::NPU::snapshot();
   py::list segments;
 
   for (const auto& segmentInfo : snapshot.segments) {
@@ -624,7 +624,7 @@ PyObject* THNPModule_memorySnapshot(PyObject* _unused, PyObject* noargs) {
   py::str oom_s = "oom";
   py::str device_free_s = "device_free";
 
-  using namespace c10_npu::NPUCachingAllocator;
+  using namespace c10_backend::CachingAllocator;
 
   auto action_to_str = [&](TraceEntry::Action action) {
     switch (action) {
@@ -703,7 +703,7 @@ PyObject* THNPModule_attachOutOfMemoryObserver(
     Py_XDECREF(result);
   };
   torch::utils::device_lazy_init(at::kPrivateUse1);
-  c10_npu::NPUCachingAllocator::attachOutOfMemoryObserver(std::move(obs));
+  c10_backend::CachingAllocator::NPU::attachOutOfMemoryObserver(std::move(obs));
   Py_RETURN_NONE;
   END_HANDLE_TH_ERRORS
 }
@@ -725,7 +725,7 @@ PyObject* THNPModule_npuCachingAllocator_raw_alloc(
   }
   ssize_t size = PyLong_AsSsize_t(size_o);
   aclrtStream stream = static_cast<aclrtStream>(PyLong_AsVoidPtr(stream_o));
-  void* mem = c10_npu::NPUCachingAllocator::raw_alloc_with_stream(size, stream);
+  void* mem = c10_backend::CachingAllocator::NPU::raw_alloc_with_stream(size, stream);
   return PyLong_FromVoidPtr(mem);
   END_HANDLE_TH_ERRORS
 }
@@ -735,14 +735,14 @@ PyObject* THNPModule_npuCachingAllocator_raw_delete(
     PyObject* obj) {
   HANDLE_TH_ERRORS
   void* mem_ptr = PyLong_AsVoidPtr(obj);
-  c10_npu::NPUCachingAllocator::raw_delete(mem_ptr);
+  c10_backend::CachingAllocator::NPU::raw_delete(mem_ptr);
   Py_RETURN_NONE;
   END_HANDLE_TH_ERRORS
 }
 
 PyObject* THNPModule_getAllocatorBackend(PyObject* _unused, PyObject* noargs) {
   HANDLE_TH_ERRORS
-  return THPUtils_packString(c10_npu::NPUCachingAllocator::name());
+  return THPUtils_packString(c10_backend::CachingAllocator::NPU::name());
   END_HANDLE_TH_ERRORS
 }
 
@@ -755,7 +755,7 @@ PyObject* THNPModule_getAllocatorBackend(PyObject* _unused, PyObject* noargs) {
 static PyGILState_STATE npuMutexGILState;
 
 PyObject* THNPModule_npuLockMutex(PyObject* module, PyObject* noargs) {
-  auto mutex = c10_npu::NPUCachingAllocator::getFreeMutex();
+  auto mutex = c10_backend::CachingAllocator::NPU::getFreeMutex();
   // This has to be a busy loop because we **absolutely need to** hold the GIL
   // or it's a recipe for a deadlock otherwise (if we let other Python threads
   // run while we have the cudaMutex, but not the GIL, they might try to e.g.
@@ -776,7 +776,7 @@ PyObject* THNPModule_npuLockMutex(PyObject* module, PyObject* noargs) {
 }
 
 PyObject* THNPModule_npuUnlockMutex(PyObject* module, PyObject* noargs) {
-  auto mutex = c10_npu::NPUCachingAllocator::getFreeMutex();
+  auto mutex = c10_backend::CachingAllocator::NPU::getFreeMutex();
   PyGILState_Release(npuMutexGILState);
   mutex->unlock();
   Py_RETURN_NONE;
